+++ pkg/server/server_authz_test.go
@@ -2,6 +2,7 @@ package server
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"testing"
 
@@ -20,6 +21,8 @@ import (
 	"github.com/openfga/openfga/pkg/tuple"
 	"github.com/openfga/openfga/pkg/typesystem"
 )
+
+var errInvalidClientID = errors.New("rpc error: code = InvalidArgument desc = client ID not found in context")
 
 type storeAndModel struct {
 	id      string
@@ -260,8 +263,8 @@ func TestListObjects(t *testing.T) {
 				Relation:             "guest",
 				User:                 "user:ben",
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_list_objects", func(t *testing.T) {
@@ -338,9 +341,8 @@ func TestStreamedListObjects(t *testing.T) {
 				Relation:             "guest",
 				User:                 "user:ben",
 			}, server)
-			require.Error(t, err)
 
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_streamed_list_objects", func(t *testing.T) {
@@ -418,8 +420,8 @@ func TestRead(t *testing.T) {
 					Object:   "workspace:1",
 				},
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_read", func(t *testing.T) {
@@ -492,8 +494,8 @@ func TestWrite(t *testing.T) {
 					},
 				},
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_write", func(t *testing.T) {
@@ -525,8 +527,8 @@ func TestWrite(t *testing.T) {
 					},
 				},
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_write_for_modules", func(t *testing.T) {
@@ -596,24 +598,31 @@ func TestCheckCreateStoreAuthz(t *testing.T) {
 
 		openfga.authorizer = authz.NewAuthorizer(&authz.Config{StoreID: settings.rootData.id, ModelID: settings.rootData.modelID}, openfga, openfga.logger)
 
+		t.Run("with_SkipAuthzCheckFromContext_set", func(t *testing.T) {
+			ctx := authclaims.ContextWithSkipAuthzCheck(context.Background(), true)
+
+			err := openfga.checkCreateStoreAuthz(ctx)
+			require.NoError(t, err)
+		})
+
 		t.Run("error_with_no_client_id_found", func(t *testing.T) {
 			err := openfga.checkCreateStoreAuthz(context.Background())
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = InvalidArgument desc = client ID not found in context", err.Error())
+
+			require.EqualError(t, err, errInvalidClientID.Error())
 		})
 
 		t.Run("error_with_empty_client_id", func(t *testing.T) {
 			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: ""})
 			err := openfga.checkCreateStoreAuthz(ctx)
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = InvalidArgument desc = client ID not found in context", err.Error())
+
+			require.EqualError(t, err, errInvalidClientID.Error())
 		})
 
 		t.Run("error_check_when_not_authorized", func(t *testing.T) {
 			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: clientID})
 			err := openfga.checkCreateStoreAuthz(ctx)
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("authz_is_valid", func(t *testing.T) {
@@ -633,7 +642,7 @@ func TestCheckAuthz(t *testing.T) {
 	ds := memory.New()
 	t.Cleanup(ds.Close)
 
-	t.Run("check_no_authz", func(t *testing.T) {
+	t.Run("authz_disabled_should_succeed", func(t *testing.T) {
 		openfga := MustNewServerWithOpts(
 			WithDatastore(ds),
 		)
@@ -645,7 +654,7 @@ func TestCheckAuthz(t *testing.T) {
 		require.NoError(t, err)
 	})
 
-	t.Run("check_with_authz", func(t *testing.T) {
+	t.Run("authz_enabled", func(t *testing.T) {
 		openfga := MustNewServerWithOpts(
 			WithDatastore(ds),
 		)
@@ -665,29 +674,29 @@ func TestCheckAuthz(t *testing.T) {
 
 		t.Run("error_with_no_client_id_found", func(t *testing.T) {
 			err := openfga.checkAuthz(context.Background(), settings.testData.id, authz.Check)
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = InvalidArgument desc = client ID not found in context", err.Error())
+
+			require.EqualError(t, err, errInvalidClientID.Error())
 		})
 
 		t.Run("error_with_empty_client_id", func(t *testing.T) {
 			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: ""})
 			err := openfga.checkAuthz(ctx, settings.testData.id, authz.Check)
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = InvalidArgument desc = client ID not found in context", err.Error())
+
+			require.EqualError(t, err, errInvalidClientID.Error())
 		})
 
 		t.Run("error_when_authorized_errors", func(t *testing.T) {
 			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: "ID"})
 			err := openfga.checkAuthz(ctx, settings.testData.id, "invalid api method")
-			require.Error(t, err)
-			require.Equal(t, authz.ErrUnknownAPIMethod.Error(), err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnknownAPIMethod)
 		})
 
 		t.Run("error_check_when_not_authorized", func(t *testing.T) {
 			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: clientID})
 			err := openfga.checkAuthz(ctx, settings.testData.id, authz.Check)
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("authz_is_valid", func(t *testing.T) {
@@ -695,6 +704,80 @@ func TestCheckAuthz(t *testing.T) {
 			settings.addAuthForRelation(ctx, t, authz.CanCallCheck)
 
 			err := openfga.checkAuthz(ctx, settings.testData.id, authz.Check)
+			require.NoError(t, err)
+		})
+	})
+}
+
+func TestGetAccessibleStores(t *testing.T) {
+	t.Cleanup(func() {
+		goleak.VerifyNone(t)
+	})
+	ds := memory.New()
+	t.Cleanup(ds.Close)
+
+	t.Run("authz_disabled_should_succeed", func(t *testing.T) {
+		openfga := MustNewServerWithOpts(
+			WithDatastore(ds),
+		)
+		t.Cleanup(openfga.Close)
+
+		_, err := openfga.getAccessibleStores(context.Background())
+		require.NoError(t, err)
+	})
+
+	t.Run("authz_enabled", func(t *testing.T) {
+		openfga := MustNewServerWithOpts(
+			WithDatastore(ds),
+		)
+		t.Cleanup(openfga.Close)
+
+		clientID := "validclientid"
+		settings := newSetupAuthzModelAndTuples(t, openfga, clientID)
+
+		openfga.authorizer = authz.NewAuthorizer(&authz.Config{StoreID: settings.rootData.id, ModelID: settings.rootData.modelID}, openfga, openfga.logger)
+
+		t.Run("with_SkipAuthzCheckFromContext_set", func(t *testing.T) {
+			ctx := authclaims.ContextWithSkipAuthzCheck(context.Background(), true)
+
+			_, err := openfga.getAccessibleStores(ctx)
+			require.NoError(t, err)
+		})
+
+		t.Run("error_with_no_client_id_found", func(t *testing.T) {
+			_, err := openfga.getAccessibleStores(context.Background())
+
+			require.EqualError(t, err, errInvalidClientID.Error())
+		})
+
+		t.Run("error_with_empty_client_id", func(t *testing.T) {
+			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: ""})
+			_, err := openfga.getAccessibleStores(ctx)
+
+			require.EqualError(t, err, errInvalidClientID.Error())
+		})
+
+		t.Run("error_when_AuthorizeListStores_errors", func(t *testing.T) {
+			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: clientID})
+			_, err := openfga.getAccessibleStores(ctx)
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
+		})
+
+		t.Run("authz_is_valid", func(t *testing.T) {
+			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: clientID})
+			_, err := settings.openfga.Write(ctx, &openfgav1.WriteRequest{
+				StoreId:              settings.rootData.id,
+				AuthorizationModelId: settings.rootData.modelID,
+				Writes: &openfgav1.WriteRequestWrites{
+					TupleKeys: []*openfgav1.TupleKey{
+						tuple.NewTupleKey("system:fga", authz.CanCallListStores, fmt.Sprintf("application:%s", settings.clientID)),
+					},
+				},
+			})
+			require.NoError(t, err)
+
+			_, err = openfga.getAccessibleStores(ctx)
 			require.NoError(t, err)
 		})
 	})
@@ -742,6 +825,13 @@ func TestCheckWriteAuthz(t *testing.T) {
 
 		openfga.authorizer = authz.NewAuthorizer(&authz.Config{StoreID: settings.rootData.id, ModelID: settings.rootData.modelID}, openfga, openfga.logger)
 
+		t.Run("with_SkipAuthzCheckFromContext_set", func(t *testing.T) {
+			ctx := authclaims.ContextWithSkipAuthzCheck(context.Background(), true)
+
+			err := openfga.checkWriteAuthz(ctx, &openfgav1.WriteRequest{}, typesys)
+			require.NoError(t, err)
+		})
+
 		t.Run("error_when_GetModulesForWriteRequest_errors", func(t *testing.T) {
 			err := openfga.checkWriteAuthz(context.Background(), &openfgav1.WriteRequest{
 				StoreId: settings.testData.id,
@@ -764,8 +854,8 @@ func TestCheckWriteAuthz(t *testing.T) {
 					},
 				},
 			}, typesys)
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("authz_is_valid", func(t *testing.T) {
@@ -793,7 +883,7 @@ func TestCheck(t *testing.T) {
 	ds := memory.New()
 	t.Cleanup(ds.Close)
 
-	t.Run("check_no_authz_should_succeed", func(t *testing.T) {
+	t.Run("authz_disabled_should_succeed", func(t *testing.T) {
 		openfga := MustNewServerWithOpts(
 			WithDatastore(ds),
 		)
@@ -816,7 +906,7 @@ func TestCheck(t *testing.T) {
 		require.True(t, checkResponse.GetAllowed())
 	})
 
-	t.Run("check_with_authz", func(t *testing.T) {
+	t.Run("authz_enabled", func(t *testing.T) {
 		openfga := MustNewServerWithOpts(
 			WithDatastore(ds),
 		)
@@ -838,8 +928,8 @@ func TestCheck(t *testing.T) {
 					Object:   fmt.Sprintf("store:%s", settings.testData.id),
 				},
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_check", func(t *testing.T) {
@@ -961,8 +1051,8 @@ func TestExpand(t *testing.T) {
 					Object:   "workspace:1",
 				},
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_expand", func(t *testing.T) {
@@ -1047,8 +1137,8 @@ func TestReadAuthorizationModel(t *testing.T) {
 					Id:      settings.testData.modelID,
 				},
 			)
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_readAuthorizationModel", func(t *testing.T) {
@@ -1116,8 +1206,8 @@ func TestReadAuthorizationModels(t *testing.T) {
 					StoreId: settings.testData.id,
 				},
 			)
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_readAuthorizationModels", func(t *testing.T) {
@@ -1204,8 +1294,8 @@ func TestWriteAssertions(t *testing.T) {
 				AuthorizationModelId: settings.testData.modelID,
 				Assertions:           assertions,
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_writeAssertions", func(t *testing.T) {
@@ -1288,8 +1378,8 @@ func TestReadAssertions(t *testing.T) {
 				StoreId:              settings.testData.id,
 				AuthorizationModelId: settings.testData.modelID,
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_readAssertions", func(t *testing.T) {
@@ -1361,8 +1451,8 @@ func TestReadChanges(t *testing.T) {
 				Type:     "user",
 				PageSize: wrapperspb.Int32(50),
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_readChanges", func(t *testing.T) {
@@ -1428,8 +1518,8 @@ func TestCreateStore(t *testing.T) {
 			_, err := openfga.CreateStore(ctx, &openfgav1.CreateStoreRequest{
 				Name: name,
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_createStore", func(t *testing.T) {
@@ -1486,8 +1576,8 @@ func TestDeleteStore(t *testing.T) {
 			_, err := openfga.DeleteStore(ctx, &openfgav1.DeleteStoreRequest{
 				StoreId: settings.testData.id,
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_deleteStore", func(t *testing.T) {
@@ -1543,8 +1633,8 @@ func TestGetStore(t *testing.T) {
 			_, err := openfga.GetStore(ctx, &openfgav1.GetStoreRequest{
 				StoreId: settings.testData.id,
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_getStore", func(t *testing.T) {
@@ -1602,8 +1692,8 @@ func TestListStores(t *testing.T) {
 				PageSize:          wrapperspb.Int32(1),
 				ContinuationToken: "",
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_listStores", func(t *testing.T) {
@@ -1685,8 +1775,8 @@ func TestListUsers(t *testing.T) {
 					{Type: "user"},
 				},
 			})
-			require.Error(t, err)
-			require.Equal(t, "rpc error: code = Code(1600) desc = the principal is not authorized to perform the action", err.Error())
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("successfully_call_listUsers", func(t *testing.T) {