+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Try to keep listed changes to a concise bulleted list of simple explanations of 
 * Goroutine leak when ListObjects or StreamedListObjects call cannot be completed within REQUEST_TIMEOUT. [#2030](https://github.com/openfga/openfga/pull/2030)
 * Label ListUsers API calls [#2000](https://github.com/openfga/openfga/pull/2000)
 * Fixed incorrect dispatch counts in List Objects [2013](https://github.com/openfga/openfga/pull/2013)
+* Fixed a bug in access control where write, create store, and list stores were not skippable [#2033](https://github.com/openfga/openfga/pull/2033)
 
 ## [1.6.2] - 2024-10-03